+++ geo/CHANGES.md
@@ -19,6 +19,8 @@
 - Fix panic in `algorithm::simplify::compute_rdp` with one point
 - Fix Clippy warning (surfaced in Rust 1.89) related to lifetime elision
 - Silence Clippy warnings related to `old_sweep` module
+- Fix false positive convexity check for "star polygon" LineStrings
+  - BREAKING: previously, an empty LineString was considered non-convex. This has changed: empty LineStrings are now considered convex, in line with tools such as PostGIS
 
 ## Update `Intersections` with new implementation of the Bentley-Ottmann sweep-line algorithm to efficiently find sparse intersections between groups of lines.